Removing `dvb-core` in favor of the in-tree version
This version will break after 4.20 due to the removal of `current_kernel_time()` anyway, and it would be a hassle.
From what I've seen, `dvb_demux.c` is pretty much the same thing, except for the kernel time fixes.
`dvb_frontend.c`, however, changed a lot, and I am unsure if it will break. It linked just fine. If it breaks, we'll have to 3-way-merge the new and old in-tree files with ours.
Until then, we're going with the in-tree `dvb-core`.